initial (partial) commit
This adds some baseline things I've found in other projects of mine that just
wind up getting copied around anyways. This adds some form of legal protection
to this copying I do (because sometimes this stuff weaves into my work commits)
and also provides a quick means of standing up new JS based projects. It is not
designed to fit every need, and it is intended that this project is copied and
the project renamed in package.json.

Some big items I have left to address are switching over to prettier.js and
removing a great deal of the validation that eslint is doing (and just let it
catch things that prettier can and perhaps should not, such as unused
variables).